Initial version of script to group files
The ZSH script processes arranges the files into groups according to
the "classifier" awk script. It is assumed that the "classifier" awk
script resides in the same directory as the "classify" script. By
default, the script works on files in the current directory. An
alternate directory may be specified as an argument. For example,
to process files in the "sample" directory:

    ./classify sample